requirements.txt
cassandra-driver==3.25.0
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
geomet==0.2.1.post1
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.3
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.3
resreg==0.2
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
Werkzeug==2.0.2
wincertstore==0.2
